feat: 재가공 컨텐츠 api 위치 수정
